Refactored sign cmd structure and dupl. code
