+++ pkg/commands/componentarchive/signature/signature.go
@@ -10,11 +10,10 @@ import (
 func NewSignatureCommand(ctx context.Context) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "signature",
-		Short: "command to work with signatures in component-descriptors",
+		Short: "command to work with signatures and digests in component-descriptors",
 	}
 
-	cmd.AddCommand(NewRSASignCommand(ctx))
-	cmd.AddCommand(NewNotarySignCommand(ctx))
+	cmd.AddCommand(NewAddDigestsCommand(ctx))
 	cmd.AddCommand(NewRSAVerifyCommand(ctx))
 	cmd.AddCommand(NewNotaryVerifyCommand(ctx))
 	cmd.AddCommand(NewCompareCommand(ctx))
